Add support for the Oak Ridge National Lab's custom Mercury schema.  References https://projects.ecoinformatics.org/ecoinfo/issues/6619.
This schema is maintained by Jim Green (jgreen@iiaweb.com) and Ranjeet Devarakonda (devarakondar@ornl.gov). The schema uses a subset of the FGDC 1998 CSDGM elements, along with custom elements specific to the Mercury Metadata web application.  Support is included in Metacat for Coordinating Node operations in the DataONE network.