Validate the session during the modification stage, rather than just assuming a CGI session (support tokens too).
Also, fix the XML document validation issue where an <additionalParty> element is added prior to the <metadataProvider> element.  This seems to be an intermittent issue, and may be due to more recent versions of perl returning hash contents more randomly than previous versions.  The %orig hash passed in to personnelList() is assumed to be random now, and I just ensured the metadataProvider is first in the produced string.

https://github.nceas.ucsb.edu/KNB/arctic-data/issues/59
refs https://github.nceas.ucsb.edu/KNB/arctic-data/issues/42