Remove hard-coded maven.home property from build.xml
The previous build.xml relied on a hard-coded maven.home ant property to run Maven which was annoying to change around as you jumped from branch to branch or system to system. The new behavior still allows a maven.home property to be set, but also falls back on $MAVEN_HOME, then falls back to asking the calling shell environment to find and print Maven home.