Don't wrap otherEntity elements in a table
This was causing really wonky behavior that Chrome couldn't parse correctly. The HTML looks fine but Chrome treats the wrapping <table> as an empty table (no rows/cells) and other parts of the output get placed in odd locatoins. The other entity types don't wrap their contents in a table so this change aligns the behavior of the otherEntity XSLT with those.