Fix breakage in BioPortalServiceTest
Jing messaged me on Slack that this test had become broken at some point and
asked for help. Jing did some debugging and noticed the response from the
BioPortal Annotator service in the test was empty and it's clear that the
application expects a result. I went over to BioPortal's web UI for the service
and confirmed the service was returning an empty result which is odd. I noticed
the test was querying the OBOE-SBC ontology which I think got absorbed by OBOE
anyway so I decided to point the test at OBOE instead of OBOE-SBC. That changed
things so the test was broken for a different reason which turned out to be an
ontology version so I changed:

  1. The test to query OBOE instead of OBOE-SBC
  2. The DatapackageSummarizer to use OBOE 1.2 instead of 1.0 in order to match
    the version of OBOE we have in BioPortal