Refactor Sitemap class to use ObjectFormatCache
Fixes #1434

In the old implementation, we used a subquery against the xml_catalog table
which was only used because of its availability and not because it was really a
canonical place to go to find a list of metadata formats. When I refactored
sitemaps in https://github.com/NCEAS/metacat/pull/1393, I didn't think to use
the ObjectFormatCache. After release, we noticed that the sitemaps on the CN
deployments were incomplete and figured out it's because not all metadata
formats are in the xml_catalog table. So we decided to make use of the
ObjectFormatCache class which is really the best place to go anyway.

In this commit, I take out the old subquery and add a method to the Sitemap
class to generate a list of values corresponding to all of the metadata format
IDs as a single-quoted, comma-separated string. Also added a basic smoke test
to the SitemapTest to make sure the the new method works.