Update OSGeo repository item in pom.xml
While doing a fresh build w/o a Maven cache, I found the build was broken
because http://download.osgeo.org/webdav/geotools is gone. After some searching
around, it appears https://repo.osgeo.org/repository/release/ is the new URL. I
can't find any official notice that this is the new official location but it
works.

Sorry for the weird whitespace in the patch. pom.xml is all over the place with
respect to whitespace so I made a best guess.