Remove # from guid.ezid.uritemplate.metadata prop
Closes #1452

We moved away from this URL scheme a few minor versions of MetacatUI ago and leaving this proprerty with the `#` symbol will result in new member nodes that fail to customize this getting URLs that don't match their MetacatUI deployment. The URLs don't break because MetacatUI handles the old links but their just not identical to what MetacatUI serves.